# Vectorize Scale16X16To8X8