@JimBobSquarePants 2021-01-21
Ah fantastic @tkp1n !!

I had a feeling this might be a bottle neck... I wonder, do you have screenshots of your trace we could use to identify further points of interest?